removed dtype argument in one_hot

+++ lstm_tf_imdb.py
@@ -298,7 +298,7 @@ def words_to_embedding(word_embedding, word_matrix):
     dim0 = word_matrix.shape[0]
     dim1 = word_matrix.shape[1]
     unrolled_matrix = tf.reshape(word_matrix,[-1])
-    one_hot = tf.one_hot(unrolled_matrix, config.vocabulary_size, dtype=tf.float32, axis=1)
+    one_hot = tf.one_hot(unrolled_matrix, config.vocabulary_size, axis=1)
     embedded_words = tf.matmul(one_hot, word_embedding)
     embedded_words = tf.reshape(embedded_words, [dim0, dim1, dim_proj])
     return embedded_words
